+++ tora/package_map.cc
@@ -25,7 +25,7 @@ void PackageMap::set(SharedPtr<Package> &pkg) {
 }
 
 void PackageMap::dump(VM *vm) {
-    printf("[dump] PackageMap(%zd)\n", data.size());
+    printf("[dump] PackageMap(%ld)\n", (long int) data.size());
     for (auto iter = data.begin(); iter != data.end(); iter++) {
         iter->second->dump(vm, 1);
     }
